requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
joblib==1.4.2
scikit-learn==1.5.1
bcrypt
flask_sqlalchemy==3.0.2
psycopg2-binary==2.9.6
alembic==1.11.1

